+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## v0.6 (Yet to release)
 
 ### Added
+- Added task delete feature on tak details modal ([issue-130](https://github.com/iluminar/goodwork/issues/130))
 - Added Global EventBus for front end notification ([issue-143](https://github.com/iluminar/goodwork/issues/143))
 - Added new permissions ('create team', 'create project', 'create office') ([issue-143](https://github.com/iluminar/goodwork/issues/143))
+- Added new permissions ('create task', 'edit task', 'delete task') ([issue-130](https://github.com/iluminar/goodwork/issues/130))
 - Added new welcome email to new registered user ([issue-26](https://github.com/iluminar/goodwork/issues/26))
 - Added new config variable `company_name` in `config/app.php` ([issue-26](https://github.com/iluminar/goodwork/issues/26))